app/requirements.txt
aerich==0.5.7
asyncpg==0.24.0
fastapi==0.68.0
pydantic==1.8.2
tortoise-orm==0.17.6
uvicorn==0.13.4
